refactor: 실시간 검색에 카운트가 제대로 증가되도록 수정
