+++ src/main/java/mju/chatuniv/statistic/domain/Statistic.java
@@ -7,6 +7,7 @@ import java.util.Map;
 import java.util.Map.Entry;
 import java.util.stream.Collectors;
 import mju.chatuniv.chat.domain.word.Word;
+import mju.chatuniv.statistic.domain.dto.StatisticResponse;
 import mju.chatuniv.statistic.exception.exceptions.StatisticNotFoundException;
 
 public class Statistic {
@@ -25,11 +26,11 @@ public class Statistic {
         words.put(word, words.getOrDefault(word, 0) + 1);
     }
 
-    public static List<Word> getWords() {
-        List<Word> result = words.entrySet()
+    public static List<StatisticResponse> getWords() {
+        List<StatisticResponse> result = words.entrySet()
                 .stream()
                 .sorted(Entry.comparingByValue(Comparator.reverseOrder()))
-                .map(Entry::getKey)
+                .map(word -> StatisticResponse.of(word.getKey().getWord(), word.getValue()))
                 .limit(10)
                 .collect(Collectors.toUnmodifiableList());
