+++ src/main/java/mju/chatuniv/statistic/service/StatisticService.java
@@ -1,15 +1,15 @@
 package mju.chatuniv.statistic.service;
 
 import java.util.List;
-import mju.chatuniv.chat.domain.word.Word;
 import mju.chatuniv.statistic.domain.Statistic;
+import mju.chatuniv.statistic.domain.dto.StatisticResponse;
 import org.springframework.scheduling.annotation.Scheduled;
 import org.springframework.stereotype.Service;
 
 @Service
 public class StatisticService {
 
-    public List<Word> findStatistics() {
+    public List<StatisticResponse> findStatistics() {
         return Statistic.getWords();
     }
